[#98] Add test for missing test_mode configs
Adds a test which removes the test_mode entry from the REP configuration
to ensure that things work as normal. Also adds a test which removes the
log_path_prefix configuration while the test_mode is configured to true
to ensure that things work as normal. These values are allowed to be
undefined in a normal audit plugin configuration.