[#169] Adjust tests to actually complete
This commit makes two changes for the tests:

1. Reload configuration before running the tests. I tried putting
this in the separate script where the REP is configured, but restarting
the server failed for some reason. This solution works and we are going
to move the configuration into the tests eventually anyway, so this
should be fine.

2. Stop the server before checking the message queue so that it
doesn't get stuck in an infinite loop of processing messages generated
by the delay server and other iRODS internal processes.